refactor: IngredientScreenViewModel 파일 위치 이동 및 코드 수정
